requirements.txt
openai
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
pyPDF2

-e .